+++ examples/flyer_chat/lib/local.dart
@@ -208,7 +208,7 @@ class LocalState extends State<Local> {
     }
   }
 
-  void _removeItem(Message item, {required int index}) async {
+  void _removeItem(Message item, {int? index, TapUpDetails? details}) async {
     await _chatController.remove(item);
   }