+++ packages/flutter_chat_ui/lib/src/utils/typedefs.dart
@@ -1,10 +1,9 @@
-import 'dart:ui';
-
+import 'package:flutter/widgets.dart';
 import 'package:flutter_chat_core/flutter_chat_core.dart';
 
 typedef ResolveUserCallback = Future<User?> Function(String id);
 typedef OnMessageTapCallback =
-    void Function(Message message, {required int index});
+    void Function(Message message, {int index, TapUpDetails details});
 typedef OnMessageSendCallback = void Function(String text);
 typedef OnAttachmentTapCallback = VoidCallback;
 typedef PaginationCallback = Future<void> Function();